+++ NEWS.md
@@ -1,3 +1,5 @@
+# rsconnect (development version)
+
 # rsconnect 1.5.0
 
 * Functions for interacting with Posit Connect deployments in