+++ NEWS.md
@@ -3,6 +3,10 @@
 * Address user registration for Posit Connect deployments hosted in Snowpark
   Container Services when there is more than one configured Snowflake
   connection. (#1189)
+
+* Process cookie expiration dates in addition to the cookie max-age. Some
+  servers return already-expired cookies. (1187)
+
 * Removed unused internal methods from Connect client. (#1182)
 
 # rsconnect 1.5.0